FastAPI_App/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.9.2
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.3.3
pytest-asyncio==0.24.0
